requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
click==8.3.1
colorama==0.4.6
fastapi==0.121.3
h11==0.16.0
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
referencing==0.37.0
rpds-py==0.29.0
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
